Added reading of sub matrices and improved build/README
* Added free-ing memory from NW_SCORING swap table after run
* Added reading from substitution pairs table (nw_load_scores.c & nw_load_scores.h)

* Fixed preset substitution matrices

* Improved Makefile to properly debug ('make DEBUG=1') / clean up ('make clean')
* Improved README to include example code for using needleman_wuncsh.c in other projects

+++ Makefile
@@ -1,9 +1,12 @@
 ifdef DEBUG
-	FLAGS="-DDEBUG=1"
+	FLAGS=-DDEBUG=1 --debug
 endif
 
 all:
-	gcc -o needleman_wunsch $(FLAGS) -Wall -lz nw_cmdline.c needleman_wunsch.c utility_lib.c string_buffer.c
+	gcc -o needleman_wunsch $(FLAGS) -Wall -lz nw_cmdline.c nw_load_scores.c needleman_wunsch.c utility_lib.c string_buffer.c
 
 clean:
-	rm needleman_wunsch
+	if test -e needleman_wunsch; then rm needleman_wunsch; fi
+	
+	for file in $(wildcard *.dSYM); do rm -r $$file; done
+	for file in $(wildcard *.greg); do rm $$file; done
